Moving more stuff to a class.
